+++ statewave/__init__.py
@@ -1,6 +1,6 @@
 """Statewave Python SDK."""
 
-__version__ = "0.10.1"
+__version__ = "0.10.2"
 
 from statewave.client import AsyncStatewaveClient, StatewaveClient, RetryConfig, DEFAULT_RETRY, NO_RETRY
 from statewave.exceptions import (